Add links to each recipe picture
The pictures fit a width of 320px to fix the smallest of devices, which
I believe is the iPhone 5. Each picture has the width set to 320px and
height set to auto to allow proper scaling of the image, no matter its
size. Specifying both a width and height could skew the image.

The link provided allows users to see the full resolution picture as
some are large.